Separate doco for native pub-sub
