--- Snippets/SqlTransport/SqlTransport_5/UpgradeGuides/4to5/ConfigureSubscriptionCache.cs
@@ -1,17 +1,0 @@
-﻿using System;
-using NServiceBus;
-using NServiceBus.Transport.SQLServer;
-
-class ConfigureSubscriptionCache
-{
-    public ConfigureSubscriptionCache(EndpointConfiguration endpointConfiguration)
-    {
-        #region 4to5-configure-subscription-cache
-
-        var transport = endpointConfiguration.UseTransport<SqlServerTransport>();
-        var subscriptions = transport.SubscriptionSettings();
-        subscriptions.CacheSubscriptionInformationFor(TimeSpan.FromMinutes(1));
-
-        #endregion
-    }
-}
